Refactor out Modal component from MobilePrompt
